refactor(api): move checkIfUserIsBlocked security pre-handler in iam context
