+++ api/src/identity-access-management/application/oidc-provider/oidc-provider.route.js
@@ -1,6 +1,6 @@
 import Joi from 'joi';
 
-import { securityPreHandlers } from '../../../shared/application/security-pre-handlers.js';
+import { checkIfUserIsBlocked } from '../security-pre-handlers.js';
 import { oidcProviderController } from './oidc-provider.controller.js';
 
 export const oidcProviderRoutes = [
@@ -120,7 +120,7 @@ export const oidcProviderRoutes = [
           }),
         }),
       },
-      pre: [{ method: securityPreHandlers.checkIfUserIsBlocked }],
+      pre: [{ method: checkIfUserIsBlocked }],
       handler: (request, h) => oidcProviderController.findUserForReconciliation(request, h),
       notes: [
         "- Cette route permet d'identifier un utilisateur Pix provenant de la double mire OIDC.\n" +
